Add files via upload
The Number Guessing Game is a fun and interactive Python project where the program randomly selects a number within a specified range, and the player tries to guess that number. This game is a popular choice for beginners as it covers essential programming concepts, such as loops, conditionals, random number generation, and user input handling. The project provides an enjoyable way to practice coding fundamentals while offering a satisfying challenge to users.